Update homescreen menu and refactor menu items into their own class
